[ISSofa] FIX: warnings with gcc 9.3
(cherry picked from commit 01f1e1ed34a6076d1e3469a060f5243efe12a8b9)

+++ modules/SofaBaseLinearSolver/EigenLLTDomainSolver.inl
@@ -203,7 +203,7 @@ void EigenLLTDomainSolver< DataTypes >::solveTriangular(const MatrixDeriv& J)
     Eigen::VectorXd Jperm;
     Jperm.resize(mstateMatrixSize);
 
-    for (int r = 0; r < J.rowIndex.size(); ++r)
+    for (int r = 0; r < (int)J.rowIndex.size(); ++r)
     {
         const int i = J.rowIndex[r];
         indexMap[r]             = i;
